Extracted Markdown converter to a separate package
