+++ .gitignore
@@ -19,3 +19,5 @@ lychee.toml
 lychee-report.md
 .phpunit.result.cache
 code_samples/_inline_php/
+/build/
+*.egg-info/
